buffer_cache: Refactor buffer manager with sparse arenas
The buffer cache is responsble for mirroring the GPU accessed address space in device local memory. On main it uses a simple scheme where the address space is divided into 16KB pages.
Buffers are allocated with that granularity and all pages of that buffer point to the same buffer object. When a range overlaps multiple buffers, a join overlap operation is performed
where a new buffer is created and the old buffer contents are copied over to the new one.

This scheme is quite simple and works well but has some issues

* It requires a 2-pass approach to buffer binding, because different bindings in the same draw/dispatch can cause a join overlap between them.
  If one binding was a storage buffer being written to, then data would be lost otherwise. This also causes various other issues when trying to batch memory uploads

* Join overlap can cause large VRAM usage spikes and requires heuristics to avoid repeated recreations.
  If the cache holds a large 1GB+ buffer, and then a range that partially overlaps it is accessed, the full 1GB range + the non overlapped size is going to get allocated.
  The old buffer will eventually be freed but this is still a large VRAM spike that can OOM or force system memory usage on lower VRAM gpus

To resolve these issues a new buffer management scheme is proposed in this pull request, that uses sparse residency buffers.
The address space is divided into 2^32 VA arenas, and each arena holds a certain amount of blocks/tiles which are determined by the sparse bind granularity of the host GPU.

The arena size has been chosen because some lower end apple sillicon GPUs only support creating buffers up to 4GB, even though the GPU hardware support much large address spaces.
Currently the code assumes that arena merges are always successfull, but that case will fail on those GPUs, for that bridge VA buffers need to be created and cached.
However accesses that straddle arenas are expected to be rare enough for this to not be an actual issue in practice.

Any buffer access is checked and overlapping blocks are made resident on demand by allocating memory. Residency requests are batched and submitted before scheduler submits the main command buffer.
If an access overlaps two adjacent arenas, then the arenas are merged into a single larger one and all their backings are brought up.

This new design avoids the old issues. There is no join operation aside from arena joining, which doesnt allocate any memory and is mostly for bookeeping purposes.
The buffer objects are also stable now so the 2-pass approach has been dropped.

For tracking residency a custom sorted vector based interval set has been added which is optimized for fast queries, as checking if a range is resident is in quite a hot path.

+++ src/video_core/texture_cache/tile_manager.h
@@ -18,16 +18,16 @@ class TileManager {
 
 public:
     using ScratchBuffer = std::pair<vk::Buffer, VmaAllocation>;
-    using Result = std::pair<vk::Buffer, u32>;
+    using Result = std::pair<vk::Buffer, u64>;
 
     explicit TileManager(const Vulkan::Instance& instance, Vulkan::Scheduler& scheduler,
                          StreamBuffer& stream_buffer);
     ~TileManager();
 
     void TileImage(Image& in_image, std::span<vk::BufferImageCopy> buffer_copies,
-                   vk::Buffer out_buffer, u32 out_offset, u32 copy_size);
+                   vk::Buffer out_buffer, u64 out_offset, u64 copy_size);
 
-    Result DetileImage(vk::Buffer in_buffer, u32 in_offset, const ImageInfo& info);
+    Result DetileImage(vk::Buffer in_buffer, u64 in_offset, const ImageInfo& info);
 
 private:
     vk::Pipeline GetTilingPipeline(const ImageInfo& info, bool is_tiler);
